feat: 프로필 이미지 추가
회원가입시 프로필 이미지 설정 및
프로필 수정시에 이미지 수정도 가능하게 추가

+++ src/components/Navbar.js
@@ -16,14 +16,18 @@ function Navbar() {
   };
 
   const logoutBtn = async () => {
-    await privateAPI
-      .post("accounts/signout/", {
-        refresh: localStorage.getItem("refreshToken"),
-      })
-      .then((response) => {
-        localStorage.clear();
-        getLogin();
-      });
+    const confirmLogout = window.confirm("로그아웃하시겠습니까?");
+
+    if (confirmLogout) {
+      await privateAPI
+        .post("accounts/signout/", {
+          refresh: localStorage.getItem("refreshToken"),
+        })
+        .then((response) => {
+          localStorage.clear();
+          window.location.reload();
+        });
+    }
   };
 
   useEffect(() => {
